Get repos with auth
Now private repos are listed as well. It's kinda hacky how you fetch the second page, even when there might not be one, and limit it to 200 repos. I'll leave it like that for now.

On another update you should consider using the search endpoint so the field can have autocomplete. It displays the user's private repos as well when sending the token with the request.